[Refactor]: 프로젝트 계층 구조 변경
