+++ src/main/java/kuit/subway/global/exception/SubwayException.java
@@ -1,4 +1,4 @@
-package kuit.subway.exception;
+package kuit.subway.global.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
