+++ src/main/java/kuit/subway/station/repository/StationRepository.java
@@ -1,6 +1,6 @@
-package kuit.subway.repository;
+package kuit.subway.station.repository;
 
-import kuit.subway.entity.Station;
+import kuit.subway.station.domain.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
